# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Gateway API
  description: Agent 网关中台后端，提供账户、项目、服务管理和模型网关接口。
  version: 0.1.0
paths:
...
  /api/admin/v1/projects/{project_id}/keys:
    get:
      tags:
      - API Keys
      summary: 查询项目 API Key
      operationId: list_api_keys_api_admin_v1_projects__project_id__keys_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKeyResponse'
                title: Response List Api Keys Api Admin V1 Projects  Project Id  Keys
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - API Keys
      summary: 创建项目 API Key
      operationId: create_api_key_api_admin_v1_projects__project_id__keys_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreatedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/v1/projects/{project_id}/keys/{key_id}:
    delete:
      tags:
      - API Keys
      summary: 撤销项目 API Key
      operationId: revoke_api_key_api_admin_v1_projects__project_id__keys__key_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Key Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ApiKeyCreateRequest:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
          description: 省略表示永不过期。
      type: object
      required:
      - name
      title: ApiKeyCreateRequest
    ApiKeyCreatedResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        project_id:
          type: string
          format: uuid
          title: Project Id
        created_by_user_id:
          type: string
          format: uuid
          title: Created By User Id
        created_by_username:
          type: string
          title: Created By Username
        name:
          type: string
          title: Name
        key_prefix:
          type: string
          title: Key Prefix
        key_last_four:
          type: string
          title: Key Last Four
        secret:
          type: string
          title: Secret
          description: 完整 Key；仅创建者之后可查看。
        status:
          type: string
          enum:
          - active
          - revoked
          - expired
          title: Status
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        last_used_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Used At
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - project_id
      - created_by_user_id
      - created_by_username
      - name
      - key_prefix
      - key_last_four
      - secret
      - status
      - expires_at
      - last_used_at
      - revoked_at
      - created_at
      title: ApiKeyCreatedResponse
    ApiKeyResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        project_id:
          type: string
          format: uuid
          title: Project Id
        created_by_user_id:
          type: string
          format: uuid
          title: Created By User Id
        created_by_username:
          type: string
          title: Created By Username
        name:
          type: string
          title: Name
        key_prefix:
          type: string
          title: Key Prefix
        key_last_four:
          type: string
          title: Key Last Four
        secret:
          anyOf:
          - type: string
          - type: 'null'
          title: Secret
          description: 完整 Key；仅 owner/editor 可查看，旧版无法恢复的 Key 为 null。
        status:
          type: string
          enum:
          - active
          - revoked
          - expired
          title: Status
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        last_used_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Used At
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - project_id
      - created_by_user_id
      - created_by_username
      - name
      - key_prefix
      - key_last_four
      - secret
      - status
      - expires_at
      - last_used_at
      - revoked_at
      - created_at
      title: ApiKeyResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer